Add all main feature

+++ term7moviemobile/lib/models/showtime_ticket_type_model.dart
@@ -2,7 +2,7 @@ import 'package:term7moviemobile/models/ticket_type_model.dart';
 
 class ShowtimeTicketTypeModel {
   String? id;
-  int? receivePrice;
+  double? receivePrice;
   TicketTypeModel? ticketType;
 
   ShowtimeTicketTypeModel(
